@@ -16,7 +16,8 @@ from pydantic import Field, field_validator, model_validator
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 if TYPE_CHECKING:
-    from typing import Any, Callable
+    from collections.abc import Callable
+    from typing import Any
 
 _DEFAULT_CONFIG_FILE_PATH = Path("~", ".quacc.yaml").expanduser().resolve()
 
@@ -136,6 +137,10 @@ class QuaccSettings(BaseSettings):
     # ---------------------------
     PREFECT_AUTO_SUBMIT: bool = Field(
         True, description="Whether to auto-submit tasks to the task runner."
+    )
+    PREFECT_RESOLVE_FLOW_RESULTS: bool = Field(
+        True,
+        description="Whether to resolve all futures in flow results to data and fail if not possible",
     )
 
     # ---------------------------
@@ -542,7 +547,7 @@ def _type_handler(settings: dict[str, Any]) -> dict[str, Any]:
 
 
 @contextmanager
-def change_settings(changes: dict[str, Any] | None):
+def change_settings(changes: dict[str, Any]):
     """
     Temporarily change an attribute of an object.
 
@@ -552,6 +557,11 @@ def change_settings(changes: dict[str, Any] | None):
         Dictionary of changes to make formatted as attribute: value.
     """
     from quacc import _internally_set_settings, get_settings
+
+    if "WORKFLOW_ENGINE" in changes:
+        raise ValueError(
+            "Cannot change the workflow engine in a context manager. Please use environment variables, e.g. `export QUACC_WORKFLOW_ENGINE=jobflow`, if you need to dynamically change the workflow engine."
+        )
 
     settings = get_settings()
     original_values = {attr: getattr(settings, attr) for attr in changes}
